refactor: - 불필요 변수 제거 - 중첩되는 불필요 components 제거 및 재사용가능하게 변경 - css 수정
