+++ src/app/(rootLayout)/layout.tsx
@@ -2,8 +2,8 @@ import { ReactNode } from "react";
 import MainHeader from "@/app/(rootLayout)/(layout)/_components/header";
 import Providers from "@/providers";
 import ToastContainer from "@/components/UI/Toast/Organisms/ToastContainer";
-import Modal from "@/components/Pages/Modal";
 import LoginContainer from "@/components/Organisms/LoginContainer";
+import { Modal } from "@/components/Pages";
 
 interface Props {
   children: ReactNode;
