+++ src/components/Pages/BrandsPage.tsx
@@ -1,17 +1,13 @@
-import { GetBrandProductsData } from "@/api/brands";
-import { Product } from "@/types/product.type";
-import ProductPrintPage from "@/components/Pages/ProductPrintPage";
+import ProductPage from "@/components/Pages/ProductPage";
 import { Suspense } from "react";
 
 interface Props {
   brandId: number;
 }
 async function BrandsPage({ brandId }: Props) {
-  const response = (await GetBrandProductsData(brandId)) as Product[];
-
   return (
     <Suspense fallback={<div>Loading...</div>}>
-      <ProductPrintPage products={response} />
+      <ProductPage brandId={brandId} />
     </Suspense>
   );
 }
