+++ src/components/Pages/index.ts
@@ -1,4 +1,4 @@
-export { default as ProductsPageContainer } from "./ProductsPageContainer";
-export { default as ProductDetailPage } from "./ProductPage";
+export { default as ProductPage } from "./ProductPage";
 export { default as BrandsPage } from "./BrandsPage";
 export { default as Modal } from "./Modal";
+export { default as ProductDetailPage } from "./ProductDetailPage";
